Updated implementation
