+++ lib/unicode.js
@@ -1,37 +1,11 @@
 'use strict';
 
-const has_icu = !!process.versions.icu;
+const has_icu = process.binding('config').hasIntl;
 const Buffer = require('buffer').Buffer;
 const kDefaultOptions = {};
 
-const icu = has_icu ? process.binding('icu') : undefined;
-exports.hasICU = has_icu;
-
 if (has_icu) {
-
-  exports.detectEncoding = function detectEncoding(source) {
-    var ret;
-    if (typeof source === 'string') {
-      ret = icu.detectEncodingString(source);
-    } else if (!(source instanceof Buffer)) {
-      throw new TypeError('source must be a string or Buffer');
-    } else {
-      ret = icu.detectEncoding(source);
-    }
-    return ret;
-  };
-
-  exports.detectEncodings = function detectEncodings(source) {
-    const m = {};
-    if (typeof source === 'string') {
-      icu.detectEncodingsString(source, m);
-    } else if ((!source instanceof Buffer)) {
-      throw new TypeError('source must be a string or Buffer');
-    } else {
-      icu.detectEncodings(source, m);
-    }
-    return m;
-  };
+  const icu = process.binding('icu');
 
   const conversions = {
     'ascii': {
@@ -41,50 +15,50 @@ if (has_icu) {
       'utf8': (source) => {
         return source;
       },
-      'ucs2': (source, options) => {
-        return icu.convertToUcs2('us-ascii', source, options);
+      'ucs2': (source) => {
+        return icu.convertToUcs2('us-ascii', source);
       }
     },
     'binary': {
-      'ascii': (source, options) => {
-        return icu.convert('us-ascii', 'iso8859-1', source, options);
+      'ascii': (source) => {
+        return icu.convert('us-ascii', 'iso8859-1', source);
       },
-      'utf8': (source, options) => {
-        return icu.convert('utf-8', 'iso8859-1', source, options);
+      'utf8': (source) => {
+        return icu.convert('utf-8', 'iso8859-1', source);
       },
-      'ucs2': (source, options) => {
-        return icu.convertToUcs2('iso8859-1', source, options);
+      'ucs2': (source) => {
+        return icu.convertToUcs2('iso8859-1', source);
       }
     },
     'utf8': {
-      'ascii': (source, options) => {
-        return icu.convert('us-ascii', 'utf-8', source, options);
+      'ascii': (source) => {
+        return icu.convert('us-ascii', 'utf-8', source);
       },
-      'binary': (source, options) => {
-        return icu.convert('iso-8859-1', 'utf-8', source, options);
+      'binary': (source) => {
+        return icu.convert('iso-8859-1', 'utf-8', source);
       },
       'ucs2': icu.convertToUcs2FromUtf8,
     },
     'ucs2': {
-      'ascii': (source, options) => {
+      'ascii': (source) => {
         if (source.length % 2 !== 0)
           throw new TypeError('Invalid UCS2 Buffer');
-        return icu.convertFromUcs2('us-ascii', source, options);
+        return icu.convertFromUcs2('us-ascii', source);
       },
-      'binary': (source, options) => {
+      'binary': (source) => {
         if (source.length % 2 !== 0)
           throw new TypeError('Invalid UCS2 Buffer');
-        return icu.convertFromUcs2('iso-8859-1', source, options);
+        return icu.convertFromUcs2('iso-8859-1', source);
       },
-      'utf8': (source, options) => {
+      'utf8': (source) => {
         if (source.length % 2 !== 0)
           throw new TypeError('Invalid UCS2 Buffer');
-        return icu.convertToUtf8FromUcs2(source, options);
+        return icu.convertToUtf8FromUcs2(source);
       }
     }
   };
 
-  exports.reencode = function convert(source, from_enc, to_enc, options) {
+  exports.transcode = function transcode(source, from_enc, to_enc, options) {
     if (!source) return;
     if (!(source.buffer instanceof ArrayBuffer))
       throw new TypeError('"source" argument must be a Buffer');
@@ -105,7 +79,7 @@ if (has_icu) {
     if (cnv_from) {
       const cnv_to = cnv_from[to_enc];
       if (cnv_to)
-        return cnv_to(source, options);
+        return cnv_to(source, options.lenient);
     }
     throw new Error(`Unsupported conversion: ${from_enc} to ${to_enc}`);
   };
@@ -168,14 +142,13 @@ if (has_icu) {
   };
 
 } else {
-  function noICU() {
-    throw new Error('This Node.js binary was built without ICU support.');
-  }
   exports.detectEncoding =
-      exports.detectEncodings =
-      exports.reencode =
-      exports.codePointAt =
-      exports.charAt = noICU;
+  exports.transcode =
+  exports.codePointAt =
+  exports.charAt =
+    function noICU() {
+      throw new Error('This Node.js binary was built without ICU support.');
+    };
 }
