+++ node.gyp
@@ -47,7 +47,7 @@
       'lib/_linklist.js',
       'lib/module.js',
       'lib/net.js',
-      'lib/icu.js',
+      'lib/unicode.js',
       'lib/os.js',
       'lib/path.js',
       'lib/process.js',
@@ -164,7 +164,6 @@
         'src/node_watchdog.cc',
         'src/node_zlib.cc',
         'src/node_i18n.cc',
-        'src/node_icu.cc',
         'src/pipe_wrap.cc',
         'src/signal_wrap.cc',
         'src/spawn_sync.cc',
@@ -203,7 +202,6 @@
         'src/node_wrap.h',
         'src/node_revert.h',
         'src/node_i18n.h',
-        'src/node_icu.h',
         'src/pipe_wrap.h',
         'src/tty_wrap.h',
         'src/tcp_wrap.h',
