--- src/node_icu.h
@@ -1,16 +1,0 @@
-#ifndef SRC_NODE_ICU_H_
-#define SRC_NODE_ICU_H_
-
-#include "node.h"
-#include "v8.h"
-
-#if defined(NODE_HAVE_I18N_SUPPORT)
-
-namespace node {
-namespace ICU {
-}  // namespace ICU
-}  // namespace node
-
-#endif  // NODE_HAVE_I18N_SUPPORT
-
-#endif  // SRC_NODE_ICU_H_
